Move large parts of postinstall to OpenProject
The addon-openproject started to bloat due to a number
of rake tasks with environment becoming a necessity.

As any of these rake tasks load the environment, which may take up to 10
seconds, we should strive to include the relevant logic into OpenProject
where possible.

The OpenProject core will expose the default functionality for setting
up relative root (and in turn, recompiling assets) as well as SCM setup.